Added unit testing for frontend and backend

+++ cheeseria-backend/src/main/java/com/cheeseria/backend/controller/CheeseController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -64,13 +65,12 @@ public class CheeseController {
         }
     }
 
-    @PostMapping("/updateCheeseById/{id}")
+    @PutMapping("/updateCheeseById/{id}")
     public ResponseEntity<Cheese> updateCheeseById(@PathVariable Long id, @RequestBody Cheese newCheeseData) {
         try {
             Optional<Cheese> oldCheeseData = cheeseRepository.findById(id);
 
             if (oldCheeseData.isPresent()) {
-                //TODO Make sure new cheese data is valid
                 Cheese updatedCheeseData = oldCheeseData.get();
                 updatedCheeseData.setName(newCheeseData.getName());
                 updatedCheeseData.setColour(newCheeseData.getColour());
@@ -89,7 +89,9 @@ public class CheeseController {
     @DeleteMapping("/deleteCheeseById/{id}")
     public ResponseEntity<Cheese> deleteCheeseById(@PathVariable Long id) {
         try {
-            //TODO Respond accordingly if no cheeses with that ID are recorded
+            if (!cheeseRepository.existsById(id)) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
             cheeseRepository.deleteById(id);
             return new ResponseEntity<>(HttpStatus.OK);
         } catch (Exception e) {
